requirements.txt
afmparser==1.0.3
alembic==1.13.3
aniso8601==9.0.1
antlr4-python3-runtime==4.13.1
astutils==0.0.6
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
ConfigArgParse==1.7
coverage==7.6.4
cryptography==43.0.3
dd==0.5.7
dnspython==2.7.0
docker==7.1.0
email_validator==2.2.0
Faker==30.8.0
flake8==7.1.1
flamapy==2.0.1
flamapy-bdd==2.0.1
flamapy-fm==2.0.1
flamapy-fw==2.0.1
flamapy-sat==2.0.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gevent==24.10.3
geventhttpclient==2.3.1
graphviz==0.20.3
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kaitaistruct==0.10
locust==2.32.0
Mako==1.3.6
MarkupSafe==3.0.2
mccabe==0.7.0
msgpack==1.1.0
networkx==3.4.2
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
ply==3.11
psutil==6.1.0
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydot==3.0.2
pyflakes==3.2.0
PyMySQL==1.1.1
pyOpenSSL==24.2.1
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.3
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-sat==1.8.dev13
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
selenium==4.25.0
selenium-wire==5.1.0
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.3
uvlparser==2.0.1
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
wsproto==1.2.0
WTForms==3.2.1
zope.event==5.0
zope.interface==7.1.0
zstandard==0.23.0

